MS44 Restructuring SpotifyAPI folder
