+++ app/src/Store/SpotifyAPI/Requests/getUserTopTracks.ts
@@ -1,19 +1,18 @@
-import { apiEndpoints } from "../Endpoints";
-import { yieldGet } from "../apiStore";
+import { apiEndpoints } from "../../Endpoints";
+import { yieldGet } from "../../apiStore";
 
-function* getAllTimeUserTopTracks() {
+function* getUserTopTracks() {
   const accessToken = localStorage.getItem("access_token");
-  const userTopTracksUrl = apiEndpoints.spotifyAllTimeUserTopTracks;
+  const userTopTracksUrl = apiEndpoints.spotifyUserTopTracks;
   try {
     // @ts-ignore
     const response = yield yieldGet(userTopTracksUrl, {Authorization: "Bearer " + accessToken,})
     // @ts-ignore
     const data = yield response.json();
-    console.log('1')
     return data;
   } catch (error) {
     console.log(error)
   }
 }
 
-export { getAllTimeUserTopTracks };
+export { getUserTopTracks };
